Address comments by @jrhee17
