+++ it/spring/boot3-athenz/build.gradle
@@ -9,13 +9,6 @@ dependencies {
     testImplementation libs.spring.boot3.starter.test
 }
 
-task generateSources(type: Copy) {
-    from("${rootProject.projectDir}/athenz/it/spring/boot3-athenz/src/test/java") {
-        exclude "**/EnableTestMetrics.java"
-    }
-    into "${project.ext.genSrcDir}/test/java"
-}
-
 // TODO(ikhoon): Consider adding `athenz-testing-internal` module.
 // Copy the Athenz Docker resources from ':athenz'.
 task copyTestResources(type: Copy) {
@@ -27,7 +20,6 @@ task copyTestResources(type: Copy) {
     into "${project.ext.genSrcDir}/test"
 }
 
-tasks.generateSources.dependsOn(tasks.copyTestResources)
 tasks.compileJava.dependsOn(tasks.copyTestResources)
 tasks.processResources.dependsOn(tasks.copyTestResources)
 tasks.processTestResources.dependsOn(tasks.copyTestResources)
